Rename to interest_area, Add reading module
Make a general InterestArea class
Make a module dedicated for reading measures.
WIP on the docstring and methods.
